Load new projector config when dataset changes
Previously, when the data provider changed, a new projector config might not load because the name of the new run (string) could stay the same - indeed the name of the run is always 'proto'.

This change makes it so that we load a new projector config and correspondingly update the UI when either an external component sets a run or we load a new data provider.